% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tufter-package.R
\docType{package}
\name{tufter-package}
\alias{tufter}
\alias{tufter-package}
\title{tufter: Implement Edward Tufte's principles of graphical design}
\description{
Use Tufte-inspired \pkg{ggplot2} layers and themes to draw figures, then
estimate their data-ink ratio, lie factor and data density.
}
\details{
I built the package to make these ideas easier to try and examine. The
measurements can help you compare drafts, but they can't tell you whether
a figure supports your argument.

\code{\link{tufte_principles}()} maps principles to functions and records
which ones can be checked. \code{\link{tufte_audit}()} runs the available
checks on an existing plot.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lobsterbush.github.io/tufter/}
  \item \url{https://github.com/lobsterbush/tufter}
  \item Report bugs at \url{https://github.com/lobsterbush/tufter/issues}
}

}
\author{
\strong{Maintainer}: Charles Crabtree \email{charles.crabtree@monash.edu} (\href{https://orcid.org/0000-0001-5144-8671}{ORCID})

}
\keyword{internal}
